fix: fix base bottom sheet
