+++ app/src/main/java/com/nexters/fullstack/widget/CreateAlbumBottomSheet.kt
@@ -1,36 +1,20 @@
 package com.nexters.fullstack.widget
 
 import android.content.Context
-import android.os.Bundle
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.bottomsheet.BottomSheetDialog
-import com.nexters.fullstack.R
+import android.util.Log
 import com.nexters.fullstack.databinding.DialogCreateLabelBinding
 
-class CreateAlbumBottomSheet(
-    context: Context
-) : BottomSheetDialog(context, R.style.BottomSheetDefaultStyle) {
+class CreateAlbumBottomSheet(context: Context) :
+    BaseBottomSheet<DialogCreateLabelBinding>(
+        viewBinding = DialogCreateLabelBinding::inflate,
+        context = context
+    ) {
 
-    lateinit var binding: DialogCreateLabelBinding
+    override fun initView() {
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DialogCreateLabelBinding.inflate(layoutInflater)
-
-        setContentView(binding.root)
-
-        initView()
-    }
-
-    override fun onStart() {
-        super.onStart()
-
-        if (behavior.state != BottomSheetBehavior.STATE_EXPANDED) {
-            behavior.state = BottomSheetBehavior.STATE_EXPANDED
+        binding.close.setOnClickListener {
+            Log.e("error", "click")
+            dismiss()
         }
-    }
-
-    private fun initView() {
-        binding.close.setOnClickListener { dismiss() }
     }
 }
