+++ app/src/main/res/drawable/bg_selected_box.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <corners android:radius="@dimen/spacing_small" />
+            <solid android:color="@color/light_gray" />
+            <size android:width="30dp" android:height="30dp" />
+        </shape>
+    </item>
+</selector>
