+++ app/src/main/res/layout/fragment_label_create.xml
@@ -11,29 +11,50 @@
         android:layout_height="match_parent">
 
         <androidx.constraintlayout.widget.ConstraintLayout
-            app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="match_parent"
+            app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior">
 
-            <EditText
+            <LinearLayout
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="@dimen/spacing_huge"
                 android:layout_marginEnd="@dimen/spacing_huge"
-                android:layout_marginBottom="@dimen/spacing_xxx_large"
                 android:background="@drawable/bg_new_label"
-                android:hint="@string/new_label_hint"
-                android:inputType="text"
-                android:maxLines="1"
-                android:paddingStart="@dimen/spacing_xxx_large"
-                android:paddingTop="@dimen/spacing_large"
-                android:paddingEnd="@dimen/spacing_xxx_large"
-                android:paddingBottom="@dimen/spacing_large"
-                android:textColorHint="@color/light_gray"
+                android:orientation="horizontal"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
+                app:layout_constraintTop_toTopOf="parent">
+
+                <ImageView
+                    android:id="@+id/select_color"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="center"
+                    android:layout_marginStart="@dimen/spacing_normal"
+                    android:src="@drawable/bg_selected_box" />
+
+                <EditText
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="center"
+                    android:background="@drawable/bg_new_label"
+                    android:hint="@string/new_label_hint"
+                    android:inputType="text"
+                    android:maxLines="1"
+                    android:paddingStart="@dimen/spacing_x_large"
+                    android:paddingTop="@dimen/spacing_large"
+                    android:paddingEnd="@dimen/spacing_x_large"
+                    android:paddingBottom="@dimen/spacing_large"
+                    android:textColorHint="@color/light_gray"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent" />
+            </LinearLayout>
+
+
         </androidx.constraintlayout.widget.ConstraintLayout>
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
 </layout>
